Struct ringbuf::ring_buffer::LocalRb

pub struct LocalRb<T, C: Container<T>> { /* private fields */ }

Ring buffer for using in single thread.

Does not implement Sync. And its Producer and Consumer do not implement Send.

This code must fail to compile:

use std::{thread, vec::Vec};
use ringbuf::LocalRb;

let (mut prod, mut cons) = LocalRb::<i32, Vec<_>>::new(256).split();
thread::spawn(move || {
    prod.push(123).unwrap();
})
.join();
thread::spawn(move || {
    assert_eq!(cons.pop().unwrap(), 123);
})
.join();

Implementations

impl<T> LocalRb<T, Vec<MaybeUninit<T>>>

pub fn new(capacity: usize) -> Self

Creates a new instance of a ring buffer.

Panics if capacity is zero.

impl<T, C: Container<T>> LocalRb<T, C>

pub unsafe fn from_raw_parts(container: C, head: usize, tail: usize) -> Self

Constructs ring buffer from container and counters.

Safety

The items in container inside head..tail range must be initialized, items outside this range must be uninitialized. head and tail values must be valid (see RbBase).

pub unsafe fn into_raw_parts(self) -> (C, usize, usize)

Destructures ring buffer into underlying container and head and tail counters.

Safety

Initialized contents of the container must be properly dropped.

pub fn split(self) -> (Producer<T, Rc<Self>>, Consumer<T, Rc<Self>>)
where Self: Sized,

Splits ring buffer into producer and consumer.

This method consumes the ring buffer and puts it on heap in Rc. If you don’t want to use heap the see Self::split_ref.

pub fn split_ref(&mut self) -> (Producer<T, &Self>, Consumer<T, &Self>)
where Self: Sized,

Splits ring buffer into producer and consumer without using the heap.

In this case producer and consumer stores a reference to the ring buffer, so you also need to store the buffer somewhere.