Struct ringbuf::ring_buffer::RbWriteCache

pub struct RbWriteCache<T, R: RbRef>
where
    R::Rb: RbWrite<T>,
{ /* private fields */ }

Caching write end of some ring buffer.

Inserted items is not visible for an opposite write end until Self::commit/Self::sync is called or Self is dropped. A free space of items removed by an opposite write end is not visible for Self until Self::sync is called.

Used to implement PostponedConsumer.

Implementations

impl<T, R: RbRef> RbWriteCache<T, R>
where R::Rb: RbWrite<T>,

pub unsafe fn new(rb_ref: R) -> Self

Create new ring buffer cache.

Safety

There must be only one instance containing the same ring buffer reference.

pub fn commit(&mut self)

Commit changes to the ring buffer.

pub fn discard(&mut self)

Discard new items pushed since last sync.

pub fn sync(&mut self)

Commit changes and fetch updates from the ring buffer.

pub fn release(self) -> R

Commit and destroy Self returning underlying ring buffer.