ringbuf::ring_buffer

Trait Container

pub unsafe trait Container<T> {
    type Internal;

    // Required methods
    fn into_internal(self) -> Self::Internal;
    unsafe fn from_internal(this: Self::Internal) -> Self;
    fn as_mut_ptr(this: &Self::Internal) -> *mut MaybeUninit<T>;
    fn len(this: &Self::Internal) -> usize;
}

Abstract container for the ring buffer.

Container items must be stored as a contiguous array.

Safety

Self::len/[Self::is_empty] must always return the same value.

Container must not cause data race on concurrent [Self::as_mut_slice]/Self::as_mut_ptr calls.

Required Associated Types

type Internal

Internal representation of the container.

Must not be aliased with its content.

Required Methods

fn into_internal(self) -> Self::Internal

Transform container to internal representation.

unsafe fn from_internal(this: Self::Internal) -> Self

Restore container from internal representation.

Safety

this must be valid.

fn as_mut_ptr(this: &Self::Internal) -> *mut MaybeUninit<T>

Return pointer to the beginning of the container items.

fn len(this: &Self::Internal) -> usize

Length of the container.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.