ringbuf::wrap::frozen

Type Alias FrozenCons

pub type FrozenCons<R> = Frozen<R, false, true>;

Frozen read end of some ring buffer.

A free space of removed items is not visible for an opposite write end until Self::commit/Self::sync is called or Self is dropped. Items inserted by an opposite write end is not visible for Self until Self::sync is called.

Aliased Type

pub struct FrozenCons<R> { /* private fields */ }