Type Alias ringbuf::consumer::PostponedConsumer

pub type PostponedConsumer<T, R> = Consumer<T, RbWrap<RbReadCache<T, R>>>;

Postponed consumer.

Aliased Type

struct PostponedConsumer<T, R> { /* private fields */ }

Implementations

impl<T, R: RbRef> PostponedConsumer<T, R>where R::Rb: RbRead<T>,

pub unsafe fn new_postponed(target: R) -> Self

Create new postponed consumer.

Safety

There must be only one consumer containing the same ring buffer reference.

pub fn sync(&mut self)

Synchronize changes with the ring buffer.

Postponed consumer requires manual synchronization to make freed space visible for the producer.

pub fn into_immediate(self) -> Consumer<T, R>

Synchronize and transform back to immediate consumer.